feat(port-forwarding): move flow status types to common
Moves the types renamed in the prior commit to common/ so that
they can be used for masquerading. Note that while the types are
moved, the uses of the types and semantics may be NAT flavor
specific. In other words, only the types are moved but not some
of the implementations.
