+++ nat/src/portfw/protocol.rs
@@ -1,76 +1,18 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Open Network Fabric Authors
 
-//! Types to represent the status of a communication in a port-forwarding context
-//! These types do not aim to represent a real state machine for any protocol.
-//! The status here modelled is aimed at determining how much the lifetime of a flow
-//! entry should be extended, or if it could be removed.
+//! Functions to represent tiny state machines for flows in the context
+//! of port forwarding. These models are very simple and aim at helping to
+//! determine how much the lifetime of flows should be extended based on
+//! the activity. This module is only for port-forwarding.
 
 use net::buffer::PacketBufferMut;
 use net::headers::TryTcp;
 use net::packet::Packet;
 use net::tcp::Tcp;
-use std::fmt::Display;
-use std::sync::Arc;
-use std::sync::atomic::AtomicU8;
 
 use super::PortFwState;
-use crate::common::NatAction;
-
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum NatFlowStatus {
-    OneWay = 0,
-    TwoWay = 1,
-    Established = 2,
-    Reset = 3,
-    CClosing = 4,
-    SClosing = 5,
-    CHalfClose = 6,
-    SHalfClose = 7,
-    LastAck = 8,
-    Closed = 9,
-}
-
-impl From<u8> for NatFlowStatus {
-    fn from(value: u8) -> Self {
-        match value {
-            0 => NatFlowStatus::OneWay,
-            1 => NatFlowStatus::TwoWay,
-            2 => NatFlowStatus::Established,
-            3 => NatFlowStatus::Reset,
-            4 => NatFlowStatus::CClosing,
-            5 => NatFlowStatus::SClosing,
-            6 => NatFlowStatus::CHalfClose,
-            7 => NatFlowStatus::SHalfClose,
-            8 => NatFlowStatus::LastAck,
-            9 => NatFlowStatus::Closed,
-            _ => unreachable!(),
-        }
-    }
-}
-impl From<NatFlowStatus> for u8 {
-    fn from(value: NatFlowStatus) -> Self {
-        value as u8
-    }
-}
-
-impl Display for NatFlowStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            NatFlowStatus::OneWay => write!(f, "oneway"),
-            NatFlowStatus::TwoWay => write!(f, "twoway"),
-            NatFlowStatus::Established => write!(f, "established"),
-            NatFlowStatus::Reset => write!(f, "reset"),
-            NatFlowStatus::CClosing => write!(f, "client-closing"),
-            NatFlowStatus::SClosing => write!(f, "server-closing"),
-            NatFlowStatus::CHalfClose => write!(f, "client-half-close"),
-            NatFlowStatus::SHalfClose => write!(f, "server-half-close"),
-            NatFlowStatus::LastAck => write!(f, "last-ack"),
-            NatFlowStatus::Closed => write!(f, "closed"),
-        }
-    }
-}
+use crate::common::{NatAction, NatFlowStatus};
 
 fn next_flow_status_tcp(pfw_state: &PortFwState, tcp: &Tcp) -> NatFlowStatus {
     let status = pfw_state.status.load();
@@ -112,8 +54,8 @@ fn next_flow_status_non_tcp(pfw_state: &PortFwState) -> NatFlowStatus {
     }
 }
 
-/// Compute the next `PortFwFlowStatus` of a flow, given the current, the received packet and
-/// the direction, which is implicit in the `PortFwAction`:
+/// Compute the next `NatFlowStatus` of a flow, given the current, the received packet and
+/// the direction, which is implicit in `PortFwState::action` (`NatAction`):
 ///     `DstNat` is the forward path and
 ///     `SrcNat` the reverse path.
 pub(crate) fn next_flow_status<Buf: PacketBufferMut>(
@@ -124,24 +66,5 @@ pub(crate) fn next_flow_status<Buf: PacketBufferMut>(
         next_flow_status_tcp(pfw_state, tcp)
     } else {
         next_flow_status_non_tcp(pfw_state)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct AtomicNatFlowStatus(Arc<AtomicU8>);
-impl AtomicNatFlowStatus {
-    #[must_use]
-    pub fn new() -> Self {
-        AtomicNatFlowStatus(Arc::new(AtomicU8::new(NatFlowStatus::OneWay.into())))
-    }
-
-    #[must_use]
-    pub fn load(&self) -> NatFlowStatus {
-        self.0.load(std::sync::atomic::Ordering::Relaxed).into()
-    }
-
-    pub fn store(&self, status: NatFlowStatus) {
-        self.0
-            .store(status.into(), std::sync::atomic::Ordering::Relaxed);
     }
 }
